DataBet/requirements.txt
amqp==5.0.6
aniso8601==9.0.1
anyio==3.1.0
appdirs==1.4.4
asgiref==3.4.1
billiard==3.6.4.0
CacheControl==0.12.6
cachetools==4.2.2
celery==5.2.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
distlib==0.3.1
Django==3.2.9
django-crontab==0.7.1
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.3
filelock==3.0.12
firebase-admin==5.0.1
Flask==2.0.1
Flask-JWT-Extended==4.2.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==0.15.1
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
h11==0.12.0
httpcore==0.13.5
httplib2==0.19.1
httpx==0.18.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.2.2
MarkupSafe==2.0.1
mongoengine==0.23.1
msgpack==1.0.2
mysql-connector-python==8.0.25
onesignal-sdk==2.0.0
packaging==21.0
pipreqs==0.4.11
prompt-toolkit==3.0.22
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==4.0.1
requests==2.25.1
rfc3986==1.5.0
rsa==4.7.2
schedule==1.1.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
virtualenv==20.4.7
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.13.3
WTForms==2.3.3
yarg==0.1.9
